web/requirements.txt
asgiref==3.5.2
Django==4.0.4
django-admin-inline-paginator==0.3.0
django-admin-ordering==0.15.0
django-cuser==2017.3.16
django-environ==0.8.1
django-js-asset==2.0.0
django-modeltranslation==0.18.2
django-tinymce==3.4.0
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.9.1
six==1.16.0
sqlparse==0.4.2